fix tests and move it to proper place
